minor stuff

+++ recurrent/random_test.go
@@ -17,7 +17,7 @@ func TestRandf(t *testing.T) {
 	t.Run("Randi() produces a random int between two others", func(t *testing.T) {
 		for i := 0; i < 100; i++ {
 			r := Randi(4, 8)
-			rem := math.Remainder(r, 1)
+			rem := math.Remainder(float64(r), 1)
 
 			if r < 1 || r > 8 {
 				t.Fail()
